add the checkmate classifier results directory
didn't upload all the input files
The results is a set of resultsX directory ( where X is a running
number)

Each result directory conatins result for checkmate classfication with
different set of features or datasets

i uploaded zip files for the smaller test sets. for the classfication.
was run on a much bigger data sets as well